Change timestamps to int64_t

+++ src/xop/G711ASource.cpp
@@ -69,9 +69,9 @@ bool G711ASource::HandleFrame(MediaChannelId channel_id, AVFrame frame)
 	return true;
 }
 
-uint32_t G711ASource::GetTimestamp()
+int64_t G711ASource::GetTimestamp()
 {
 	auto time_point = chrono::time_point_cast<chrono::microseconds>(chrono::steady_clock::now());
-	return (uint32_t)((time_point.time_since_epoch().count()+500)/1000*8);
+	return (int64_t)((time_point.time_since_epoch().count()+500)/1000*8);
 }
